created `AuthorRepository.cs` and `AuthorService.cs`
