+++ src/Chirp.Infrastructure/CheepRepository.cs
@@ -17,40 +17,6 @@ public class CheepRepository :  ICheepRepository
         this._dbContext = dbContext;
     }
 
-    public async Task<List<Author>> GetAuthor(string identifier)
-    {
-        if (identifier.Contains("@"))
-        {
-            return await GetAuthorByEmail(identifier);
-        }
-        return await GetAuthorByUserName(identifier);
-    }
-
-    public async Task<List<Author>> GetAuthorByUserName(string username)
-    {
-        var query = (from author in _dbContext.Authors
-            where author.UserName == username
-            orderby author.DisplayName
-            select author);
-        return await query.ToListAsync();
-    }
-
-    public async Task<List<Author>> GetAuthorByEmail(string email)
-    {
-        var query = (from author in _dbContext.Authors
-            where author.Email == email
-            orderby author.DisplayName
-            select author);
-        return await query.ToListAsync();
-    }
-
-    public async Task CreateAuthor(string name, string email)
-    {
-        var author = Author.Create(name, email);
-        await _dbContext.Authors.AddAsync(author);
-        await _dbContext.SaveChangesAsync();
-    }
-
     public async Task CreateCheep(Author author, string message, DateTime timestamp)
     {
         if (message.Length > Cheep.MAX_TEXT_LENGTH) {
@@ -91,60 +57,4 @@ public class CheepRepository :  ICheepRepository
         return await query.ToListAsync();
     }
 
-    public async Task Follow(Author follower, Author followed)
-    {
-        if (await ValidifyfollowRelationAsync(follower, followed))
-        {
-            return;
-        }
-        FollowRelation newFollowRelation = new FollowRelation() { Follower = follower, Followed = followed };
-        await _dbContext.AddAsync(newFollowRelation);
-        _dbContext.SaveChanges();
-    }
-
-    public async Task Unfollow(Author followerToDelete, Author followedToDelete)
-    {
-        FollowRelation followRelationToDelete = (from followRelation in _dbContext.FollowRelations
-        where followRelation.Follower == followerToDelete && followRelation.Followed == followedToDelete
-        select followRelation).First();
-        if (followedToDelete == null)
-        {
-            return;
-        }
-        _dbContext.FollowRelations.Remove(followRelationToDelete);
-        _dbContext.SaveChanges();
-    }
-    /**
-     * Returns true if breaks rules
-     */
-    private async Task<bool> ValidifyfollowRelationAsync(Author follower, Author followed)
-    {
-        if (!_dbContext.Authors.Any(author => author == follower)||
-            !_dbContext.Authors.Any(author => author == followed)||
-            follower.Id == followed.Id|| 
-            (await Following(follower)).Contains(followed)) //checks if follower already follows followed :3
-        { 
-            return true;
-        }
-        return false;   
-    }
-    /**
-     * returns all FollowRelations where `author` is follower 
-     */
-    public async Task<List<FollowRelation>> GetFollowRelations(Author author)
-    {
-        return await (from followRelation in _dbContext.FollowRelations
-        where followRelation.Follower == author
-        select followRelation).ToListAsync();
-    }
-
-    /**
-     * this is borderline unreadable, but it just gets all Authors which `author` follows
-     */
-    public async Task<List<Author>> Following(Author author)
-    {
-        return(from user in _dbContext.FollowRelations
-                where user.Follower.UserName == author.UserName
-                select user.Followed).ToList();
-    }
 }
