+++ src/Chirp.Infrastructure/CheepService.cs
@@ -29,10 +29,6 @@ public class CheepService : ICheepService
         return await cheepRepository.GetCheepsFromUserName(username, pageNr);
     }
 
-    public async Task<List<Author>> GetAuthorByUserName(string username)
-    {
-        return await cheepRepository.GetAuthorByUserName(username);
-    }
     public async Task CreateCheep(Author author, string message)
     {
         DateTime date = DateTime.Now;
