+++ src/Chirp.Infrastructure/IAuthorService.cs
@@ -1,0 +1,9 @@
+using Chirp.Core;
+using Chirp.Core.Domain_Model;
+
+namespace Chirp.Infrastructure;
+
+public interface IAuthorService
+{
+    public Task<List<Author>> GetAuthorByUserName(string username);
+}
