+++ src/Chirp.Infrastructure/ICheepService.cs
@@ -7,6 +7,5 @@ public interface ICheepService
 {
     public Task<List<CheepDTO>> GetCheeps(int pageNr);
     public Task<List<CheepDTO>> GetCheepsFromUserName(string username, int pageNr);
-    public Task<List<Author>> GetAuthorByUserName(string username);
     public Task CreateCheep(Author author, string message);
 }
