+++ src/Chirp.Web/Pages/Public.cshtml.cs
@@ -5,7 +5,7 @@ namespace Chirp.Web.Pages;
 
 public class PublicModel : CheepTimelineModel
 {
-    public PublicModel(ICheepService service) : base(service)
+    public PublicModel(ICheepService cheepService, IAuthorService authorService) : base(cheepService, authorService)
     {
     }
 
@@ -13,7 +13,7 @@ public class PublicModel : CheepTimelineModel
     {
         int pageNr = getPageNr(Request);
 
-        Cheeps = await _service.GetCheeps(pageNr);
+        Cheeps = await _cheepService.GetCheeps(pageNr);
         return Page();
     }
 }
